feat(mgs1-codec): add text-based static background
Add elements of the background that require fonts. This includes the GDQ
title at the top of the page, the arrows on the buttons, and the
background donation total display.

As an added bonus, the dynamic foreground display of the donation total is
present, but none of its functionality has been developed at this time.

A README has been added to give credit where it is required. It has been
updated to credit the creator of the DSEG7 Classic Mini font which is
used to display the donation total.