chore(mgs1-codec): correct gutters in sprite sheet
The sprite sheet used to have inconsistent gutters.  This may have been
intentional, but without knowing why, it is difficult to work with. I
have modified the sheet so within each character section, there is an 8
pixel horizontal gutter and an 11 pixel vertical gutter.

Additionally, eyes and mouth specific sprites have been moved to be in
common positions for most sprites. To be specific, 0 is the original,
1/2 are eye sprites, and 3/4 are mouth sprites.