feat(mgs1-codec): add reducer state machine
The previous version of this project was becoming too large to manage
its state.  To get around this, most of the state has been consolidated
into a reducer instead.

As a part of this, there is now a way to load scripts and read them line
by line.  This allows everything to be synchronized from the top level
of the react component.