feat(mgs1-codec): implement conversation switcher
This allows multiple conversations to be played in succession. It parses
them in order with the current implementation. Additionally, the current
setup does not handle the component loading and unloading. One side
effect of this is that the codec channel will not work properly after
switching to another scene.

Future updates should build this into the react state instead of storing
a regular javascript variable.