fix(mgs1-codec): fix component state management
This change makes the conversation part of the state itself. This allows
the component to properly reset the state when it is unloaded and
reloaded. The current behavior is to move on to the next conversation
during one of these shifts.

Additionally, the PortraitComponent had it's setTimeout management fixed
so it is guaranteed to be cleaned up.